+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/BulkExecutor.java
@@ -818,6 +818,7 @@ public final class BulkExecutor<TContext> implements Disposable {
         RequestOptions options = new RequestOptions();
         options.setThroughputControlGroupName(cosmosBulkExecutionOptions.getThroughputControlGroupName());
         options.setExcludedRegions(cosmosBulkExecutionOptions.getExcludedRegions());
+        options.setKeywordIdentifiers(cosmosBulkExecutionOptions.getKeywordIdentifiers());
 
         //  This logic is to handle custom bulk options which can be passed through encryption or through some other project
         Map<String, String> customOptions = cosmosBulkExecutionOptions.getHeaders();