+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/query/DocumentQueryExecutionContextBase.java
@@ -137,6 +137,8 @@ implements IDocumentQueryExecutionContext<T> {
         }
         request.requestContext.setExcludeRegions(
             qryOptAccessor.getImpl(cosmosQueryRequestOptions).getExcludedRegions());
+        request.requestContext.setKeywordIdentifiers(
+            qryOptAccessor.getImpl(cosmosQueryRequestOptions).getKeywordIdentifiers());
 
         request.requestContext.setIsRequestCancelledOnTimeout(this.isQueryCancelledOnTimeout);
         return request;
@@ -332,6 +334,7 @@ implements IDocumentQueryExecutionContext<T> {
             if (endToEndOperationLatencyConfig != null) {
                 executeQueryRequest.requestContext.setEndToEndOperationLatencyPolicyConfig(endToEndOperationLatencyConfig);
             }
+            executeQueryRequest.requestContext.setKeywordIdentifiers(qryOptAccessor.getImpl(cosmosQueryRequestOptions).getKeywordIdentifiers());
 
             executeQueryRequest.setPartitionKeyDefinition(qryOptAccessor.getPartitionKeyDefinition(this.cosmosQueryRequestOptions));