+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosReadManyRequestOptions.java
@@ -13,6 +13,7 @@ import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
 
 import java.time.Duration;
 import java.util.List;
+import java.util.Set;
 
 /**
  * Specifies the options associated with read many operation
@@ -309,6 +310,26 @@ public final class CosmosReadManyRequestOptions {
         this.actualRequestOptions.setCustomItemSerializer(customItemSerializer);
 
         return this;
+    }
+
+    /**
+     * Sets the custom ids.
+     *
+     * @param keywordIdentifiers the custom ids.
+     * @return the current request options.
+     */
+    public CosmosReadManyRequestOptions setKeywordIdentifiers(Set<String> keywordIdentifiers) {
+        this.actualRequestOptions.setKeywordIdentifiers(keywordIdentifiers);
+        return this;
+    }
+
+    /**
+     * Gets the custom ids.
+     *
+     * @return the custom ids.
+     */
+    public Set<String> getKeywordIdentifiers() {
+        return this.actualRequestOptions.getKeywordIdentifiers();
     }
 
     CosmosQueryRequestOptionsBase<?> getImpl() {